+++ src/ui/gtk_ui.rs
@@ -101,7 +101,7 @@ impl GtkUi {
                     let state = cell.get()
                         .expect("Thread local once cell should be initialized by now");
                     if let Some(ref event_bus) = state.event_bus {
-                        event_bus.access().emit(crate::events::Event::SwitchOff);
+                        event_bus.emit(crate::events::Event::SwitchOff);
                     }
                 })
             }));
@@ -187,7 +187,7 @@ impl GtkUi {
 
             if matches!(event, KeyEvent::Pressed) && character == 'R' {
                 state.event_bus.as_ref().map(|event_bus| {
-                    event_bus.access().emit(crate::events::Event::Reset);
+                    event_bus.emit(crate::events::Event::Reset);
                 });
             }